% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOptimalTuning.R
\name{findOptimalTuning}
\alias{findOptimalTuning}
\title{Tune Random Rotation Forest parameters.}
\usage{
findOptimalTuning(x, y, k = 10, paraGrid, verboset = TRUE, ...)
}
\arguments{
\item{x}{Training data input matrix.}

\item{y}{Training data response.}

\item{k}{Number of folds.}

\item{paraGrid}{Parameter grid to search over for optimal parameters.}

\item{verboset}{Boolean, if true prints progress, otherwise prints nothing.}

\item{...}{Additional arguments specified to \code{RRotF}.}
}
\value{
A list containing the optimal tuning parameters and CV errors per fold.
}
\description{
Finds optimal tuning parameters using k-fold cross-validation.
}
\author{
Arnu Pretorius <arnupretorius@gmail.com>
}
\seealso{
\code{\link{kFoldRun}}
}

