% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kFoldRun.R
\name{kFoldRun}
\alias{kFoldRun}
\title{k-fold cross-validiation for Random Rotation Forests.}
\usage{
kFoldRun(x, y, k = 10, seed = 1, verbose = TRUE, ...)
}
\arguments{
\item{x}{Training data input matrix.}

\item{y}{Training data response.}

\item{k}{Number of folds.}

\item{seed}{Specified seed for reproducible folds.}

\item{verbose}{Boolean, if true prints progress of k-fold CV runs, if false prints nothing.}

\item{...}{Additional arguments specified to \code{RRotF}.}
}
\value{
A list containing the average CV error computed over the folds and the error per fold.
}
\description{
Performs k-fold cross-validation using specified Random Rotation Forest as prediction model.
}
\examples{
library(ElemStatLearn)
library(caret)
data("SAheart")
trainIndex <- createDataPartition(SAheart$chd, p=0.6, list=FALSE)
train <- SAheart[trainIndex,]
test <- SAheart[-trainIndex,]
Xtrain <- train[,-10]
ytrain <- train[,10]
Xtest <- test[,-10]
ytest <- test[,10]
CVerror <- kFoldRun(Xtrain, ytrain, k=5)
CVerror
}
\author{
Arnu Pretorius <arnupretorius@gmail.com>
}

