% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictRRotF.R
\name{predict.RRotF}
\alias{predict.RRotF}
\title{Predict method for RRotF objects.}
\usage{
\method{predict}{RRotF}(object, newdata, type = "class")
}
\arguments{
\item{object}{An object of class \code{RRotF}, as created by the function \code{RRotF}.}

\item{newdata}{A data frame with the same predictors as in the training data.}

\item{type}{Type of prediction: 'prob' for probabilities and 'class' for class labels.}
}
\value{
A vector containing either predicted class probabilities or labels.
}
\description{
Prediction of new observations using RRotF model.
}
\examples{
library(ElemStatLearn)
library(caret)
data("SAheart")
trainIndex <- createDataPartition(SAheart$chd, p=0.6, list=FALSE)
train <- SAheart[trainIndex,]
test <- SAheart[-trainIndex,]
Xtrain <- train[,-10]
ytrain <- train[,10]
Xtest <- test[,-10]
ytest <- test[,10]
mod <- RRotF(Xtrain, ytrain, model="log")
preds <- predict(mod, Xtest)
error <- mean(preds != ytest)
error
}
\author{
Arnu Pretorius: \email{arnupretorius@gmail.com}
}
\references{
Rodriguez, J.J., Kuncheva, L.I., 2006. Rotation forest: A new classifier ensemble method. IEEE Trans. Pattern Anal. Mach. Intell. 28, 1619-1630. doi:10.1109/TPAMI.2006.211
}
\seealso{
\code{\link{RRotF}}
}

